README
##### FOR EDUCATIONAL PURPOSES ONLY ####

This is a simple DuckyScript that has been tested on the O.MG cable to circumvent windows Defender by adding an exception rule to a created folder. This also runs the cmd prompt and PowerShell app as administrator as well, providing the current logged in target user has elevated permissions to run cmd as admin.